Support for dependent consumers. There are use-cases where multiple consumer are consuming off same topic and these consumer have parent/child relationship. i.e. A < B < C Consumer A need to be always behind B and so B behind C in consumption.
One of the approach is to create another topic for A and when B read it writes to another top which A feeds from. This will lead to unnecessary redundant copy of same data.

I have implemented another implementation of ParitionManager -> DependentConsumerPartitionManager .
 It checks the current offset of ParentConsumer (specified as depConsumerId in DependentConsumerSpoutConfig) and prevents current consumer from going ahead of it.

For performance reason, parent consumer is fetched every 2 seconds (configurable).

As per usage, KafkaSpout need to be initialised with DependentConsumerSpoutConfig instead of SpoutConfig

Current Limitation:
- Dependent consumer need to be a SimpleConsumer and tracked in same ZK.